% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{cc_colors}
\alias{CC_COLORS}
\alias{CC_COLOURS}
\alias{cc_colors}
\alias{cc_colours}
\alias{ychrom_colors}
\title{Canonical colour scheme for the founder strains of the Collaborative Cross and Diversity Outbred}
\usage{
cc_colors(...)

CC_COLORS(...)

ychrom_colors(...)

cc_colours(...)

CC_COLOURS(...)
}
\arguments{
\item{...}{ignored}
}
\value{
a named, vector of hex colors, indexed by full strain names

a named, vector of hex colors, indexed by one-letter strain codes

Color palette for Y chromosome haplogroups in CC and DO
}
\description{
Canonical colour scheme for the founder strains of the Collaborative Cross and Diversity Outbred
}
\seealso{
\code{\link{cc_strains}}
}

