% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{cc_diplotypes}
\alias{cc_diplotypes}
\title{Return CC diplotypes in canonical order}
\usage{
cc_diplotypes(strains = LETTERS[1:8], phased = FALSE, ...)
}
\arguments{
\item{strains}{which strains to include in diplotype states; default is all 8}

\item{phased}{logical; if FALSE (the default), return unordered combinations; otherwise respect order}

\item{...}{ignored}
}
\value{
a character vector of diplotype states as 2-letter codes
}
\description{
Return CC diplotypes in canonical order
}

