% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromsizes.R
\name{chromnames}
\alias{chromnames}
\title{Just return chromosome names.}
\usage{
chromnames(species = c("mouse", "dog", "human"), ...)
}
\arguments{
\item{species}{return chromosome names for this species (default is mouse)}

\item{...}{ignored}
}
\value{
character vector of chromosome names in the UCSC style: chr1, ..., chrX, chrY, chrM
}
\description{
Just return chromosome names.
}

