% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromsizes.R
\name{chromsizes_cM}
\alias{chromsizes_cM}
\title{Chromosome lengths in cM.}
\usage{
chromsizes_cM(...)
}
\arguments{
\item{...}{ignored}
}
\value{
vector of chromosome lenghs in cM
}
\description{
Chromosome lengths in cM.
}
\details{
Lengths are from Collaborative Cross G2:F1 recombination map.
}

