% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromsizes.R
\name{chromsizes_mm10}
\alias{chromsizes_mm10}
\title{Shortcut to chromosome lengths for genome build mm10.}
\usage{
chromsizes_mm10(...)
}
\arguments{
\item{...}{passed through to \code{chromsizes()}}
}
\value{
either a named vector of chromosome lengths or a \code{Seqinfo} object; see \code{?chromsizes}
}
\description{
Shortcut to chromosome lengths for genome build mm10.
}
\details{
Chromosome names are in the "chr1", ..., "chrM" (UCSC) style.  If you prefer Ensembl ("1", ..., "MT) style,
	use \code{chromsizes("38")}.
}
\seealso{
\code{\link{chromsizes}}
}

