% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromsizes.R
\name{chromsizes_mm9}
\alias{chromsizes_mm9}
\title{Shortcut to chromosome lengths for genome build mm9.}
\usage{
chromsizes_mm9(...)
}
\arguments{
\item{...}{passed through to \code{chromsizes()}}
}
\value{
either a named vector of chromosome lengths or a \code{Seqinfo} object; see \code{?chromsizes}
}
\description{
Shortcut to chromosome lengths for genome build mm9.
}
\details{
Chromosome names are in the "chr1", ..., "chrM" (UCSC) style.  If you prefer Ensembl ("1", ..., "MT) style,
	use \code{chromsizes("37")}.
	PS: Should you really still be using mm9?  Consider updating to the post-2011 genome assembly.
}
\seealso{
\code{\link{chromsizes}}
}

