% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromsizes.R
\name{factor_chrom}
\alias{factor_chrom}
\title{Create a factor whose levels are chromosome names.}
\usage{
factor_chrom(x, species = c("mouse", "dog", "human"), ...)
}
\arguments{
\item{x}{a vector of values}

\item{species}{return chromosome names for this species (default is mouse)}

\item{...}{ignored}
}
\value{
factor whose levels are chromosome names
}
\description{
Create a factor whose levels are chromosome names.
}

