% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromsizes.R
\name{factor_sex}
\alias{factor_sex}
\title{Classify sex in systematic way}
\usage{
factor_sex(x, ...)
}
\arguments{
\item{x}{a vector of sexes, encoded in any of the acceptable ways (see Details)}

\item{...}{ignored}
}
\value{
a factor with one level for each sex
}
\description{
Classify sex in systematic way
}

