% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{is_outgroup}
\alias{is_outgroup}
\title{Test if the listed taxa are outgroups with respect to Mus musculus}
\usage{
is_outgroup(x, ..)
}
\arguments{
\item{x}{a vector of values to test (will be coerced to character)}
}
\value{
a logical vector
}
\description{
Test if the listed taxa are outgroups with respect to Mus musculus
}

