% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{legend_inside}
\alias{legend_inside}
\title{Move the legend inside plotting area}
\usage{
legend_inside(where = c("bottomright", "bottomleft", "topleft", "topright"),
  ...)
}
\description{
Move the legend inside plotting area
}

