% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{linearize_genome}
\alias{linearize_genome}
\title{Lay chromosomes out end-to-end and re-calculate position of genomic intervals/points}
\usage{
linearize_genome(df, chrlen = NULL, space = 7, ...)
}
\description{
Lay chromosomes out end-to-end and re-calculate position of genomic intervals/points
}

