% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{mus_colors}
\alias{mus_colors}
\alias{mus_colours}
\title{Canonical colour scheme for Mus taxa}
\usage{
mus_colors(...)

mus_colours(...)
}
\arguments{
\item{...}{ignored}
}
\value{
a named vector of colours
}
\description{
Canonical colour scheme for Mus taxa
}

