% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{mus_taxa}
\alias{mus_taxa}
\title{Get a list of taxa in the genus Mus}
\usage{
mus_taxa(...)
}
\arguments{
\item{...}{ignored}
}
\value{
a vector of taxa names or abbreviations; these are the last piece in the Linnean binomial (for full species)
	or trinomial (for subspecies) nomenclature
}
\description{
Get a list of taxa in the genus Mus
}
\details{
For the three house mouse subspecies, abbreviations ("dom","mus","cas") are also included.
}

