% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{palettes}
\alias{palettes}
\alias{scale_color_CC}
\alias{scale_color_cc}
\alias{scale_color_chromtype}
\alias{scale_color_mus}
\alias{scale_color_sex}
\alias{scale_color_ychrom}
\alias{scale_colour_cc}
\alias{scale_colour_chromtype}
\alias{scale_colour_mus}
\alias{scale_colour_sex}
\alias{scale_colour_ychrom}
\alias{scale_fill_CC}
\alias{scale_fill_cc}
\alias{scale_fill_chromtype}
\alias{scale_fill_mus}
\alias{scale_fill_sex}
\alias{scale_fill_ychrom}
\alias{scale_shape_sex}
\title{Color palettes related to mouse genetics for use with \code{ggplot2}.}
\usage{
scale_color_cc(..., na.value = "grey", misspell = FALSE)

scale_color_CC(..., na.value = "grey")

scale_fill_cc(..., misspell = FALSE, na.value = "grey")

scale_fill_CC(..., na.value = "grey")

scale_color_ychrom(..., na.value = "grey")

scale_colour_ychrom(...)

scale_fill_ychrom(..., na.value = "grey")

scale_color_mus(..., na.value = "grey")

scale_fill_mus(..., na.value = "grey")

scale_color_sex(..., na.value = "grey", lighter = FALSE)

scale_fill_sex(..., na.value = "grey", lighter = FALSE)

scale_shape_sex(..., na.value = 18)

scale_colour_cc(...)

scale_colour_mus(...)

scale_colour_sex(...)

scale_color_chromtype(..., include.Y = FALSE)

scale_colour_chromtype(...)

scale_fill_chromtype(..., include.Y = FALSE)
}
\arguments{
\item{...}{passed through to underlying \code{ggplot2} functions}

\item{misspell}{logical; if \code{TRUE}, use common misspelling for name of strain 'NZO'}
}
\description{
Color palettes related to mouse genetics for use with \code{ggplot2}.
}
\details{
Palettes with lowercase names ('cc') are indexed by full strain names; those indexed by uppercase names ('CC'), by strain codes.
	Both one-letter and two-letter strain codes are supported.
	
	The default behavior of \code{ggplot2} scales is to remove from the legend all factor levels not present in the data, and to hide from the plot
	any values which are not represented in the scale (which are coerced to \code{NA}s.)  Use \code{na.value} to override this behavior.
	
	All \code{scale_colour_*} functions are just aliases to their \code{scale_color_*} counterparts, to be friendly to UK-style spellers.
}
\seealso{
\code{\link[ggplot2]{scale_manual}}
}

