% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromsizes.R
\name{pseudoautosomal_boundary}
\alias{PAR_mm10}
\alias{PAR_mm9}
\alias{pseudoautosomal_boundary}
\alias{scale_x_Mb}
\alias{scale_y_Mb}
\title{Return position of pseudoautosomal boundary on X chromosome}
\usage{
pseudoautosomal_boundary(build = c("mm9", "mm10", "37", "38", "NCBIm37",
  "GRCm38"), ...)

PAR_mm9(...)

PAR_mm10(...)

scale_x_Mb(name = "position (Mb)", ...)

scale_y_Mb(name = "position (Mb)", ...)
}
\arguments{
\item{build}{genome build: see \code{?chromsizes} for options}

\item{...}{ignored}
}
\value{
a vector of length two, with the canonical boundary of pseudoautosomal region (PAR) as well as its shifted boundary
}
\description{
Return position of pseudoautosomal boundary on X chromosome
}
\references{
White MW, Ikeda A, Payseur BA (2011) A pronounced evolutionary shift of the pseudoautosomal region boundary in house mice. Mammalian Genome 23: 455-466.
	doi:10.1007/s00335-012-9403-5
}

