% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_x_linearized_genome}
\alias{scale_x_linearized_genome}
\title{Draw x-axis for concatenated chromosomes}
\usage{
scale_x_linearized_genome(..., chrlen = NULL, space = 7)
}
\description{
Draw x-axis for concatenated chromosomes
}

