% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{sex_colors_lighter}
\alias{sex_colors_lighter}
\title{Lighter colour scheme for denoting male/female sex}
\usage{
sex_colors_lighter(...)
}
\arguments{
\item{...}{ignored}
}
\value{
a named vector of colours
}
\description{
Lighter colour scheme for denoting male/female sex
}
\details{
several different notations for sex are covered here: "f"/"m", "F"/"M", "femalee"/"male", "XX"/"XY"
}

