% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{theme_classic2}
\alias{theme_classic2}
\title{The "classic" scheme of `ggplot2`, with axis-line issue fixed and nicer facet labels}
\usage{
theme_classic2(...)
}
\description{
The "classic" scheme of `ggplot2`, with axis-line issue fixed and nicer facet labels
}

