% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{chrom_colors}
\alias{chrom_colors}
\title{Canonical colour scheme for denoting autosomes vs X chromosome}
\usage{
chrom_colors(include.Y = FALSE, ...)
}
\arguments{
\item{include.Y}{logical; if \code{TRUE}, give Y chrom its own color}

\item{...}{ignored}
}
\value{
a named vector of colours
}
\description{
Canonical colour scheme for denoting autosomes vs X chromosome
}

