% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromsizes.R
\name{factor_chromtype}
\alias{factor_chromtype}
\title{Classify chromosmes as X or autosome}
\usage{
factor_chromtype(x, ...)
}
\arguments{
\item{x}{a vector of values}

\item{...}{ignored}
}
\value{
factor whose levels are "X" (X chromosome) "A" (autosome)
}
\description{
Classify chromosmes as X or autosome
}
\details{
Chromsomes which are not X or autosome will be set to NA.
}

